Add a node.js-friendly index-file
