+++ app.js
@@ -1,18 +1,7 @@
 'use strict';
 
 const express = require('express');
-const _ = require('lodash');
-
-const data = require('./data.json');
-
-function makeHorse() {
-    return {
-        name: _.sample(data.names),
-        mood: _.sample(data.moods),
-        talents: _.shuffle(data.talents).slice(0, _.sample([1,2,2,2,2,3])),
-        id: (Math.random() + '').slice(2),
-    };
-}
+const { makeHorse } = require('./index');
 
 var app = express();
